compose: Add Compose table dump internal API
- Move `print_compose_table_entry` to own file and add a `file`
  argument to select output.
- Add `xkb_compose_table_dump` to dump a Compose table.

This change is needed in order to share the feature “dump a Compose
table” between tests and tools.
